+++ src/seekstorm_server/api_endpoints.rs
@@ -18,8 +18,8 @@ use seekstorm::{
     index::{
         AccessType, DeleteDocument, DeleteDocuments, DeleteDocumentsByQuery, DistanceField,
         Document, Facet, FileType, IndexArc, IndexDocument, IndexDocuments, IndexMetaObject,
-        MinMaxFieldJson, SchemaField, SimilarityType, Synonym, TokenizerType, UpdateDocument,
-        UpdateDocuments, create_index, open_index,
+        MinMaxFieldJson, SchemaField, SimilarityType, StemmerType, Synonym, TokenizerType,
+        UpdateDocument, UpdateDocuments, create_index, open_index,
     },
     ingest::IndexPdfBytes,
     search::{FacetFilter, QueryFacet, QueryType, ResultSort, ResultType, Search},
@@ -127,6 +127,8 @@ pub struct CreateIndexRequest {
     pub similarity: SimilarityType,
     #[serde(default = "tokenizer_type_api")]
     pub tokenizer: TokenizerType,
+    #[serde(default = "stemmer_type_api")]
+    pub stemmer: StemmerType,
     #[schema(required = true, example = json!([{"terms":["berry","lingonberry","blueberry","gooseberry"],"multiway":false}]))]
     #[serde(default)]
     pub synonyms: Vec<Synonym>,
@@ -138,6 +140,10 @@ fn similarity_type_api() -> SimilarityType {
 
 fn tokenizer_type_api() -> TokenizerType {
     TokenizerType::UnicodeAlphanumeric
+}
+
+fn stemmer_type_api() -> StemmerType {
+    StemmerType::None
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -407,12 +413,14 @@ pub(crate) async fn open_all_apikeys(
         (status = UNAUTHORIZED, description = "API key does not exists")
     )
 )]
+#[allow(clippy::too_many_arguments)]
 pub(crate) fn create_index_api<'a>(
     index_path: &'a PathBuf,
     index_name: String,
     schema: Vec<SchemaField>,
     similarity: SimilarityType,
     tokenizer: TokenizerType,
+    stemmer: StemmerType,
     synonyms: Vec<Synonym>,
     apikey_object: &'a mut ApikeyObject,
 ) -> u64 {
@@ -435,6 +443,7 @@ pub(crate) fn create_index_api<'a>(
         name: index_name,
         similarity,
         tokenizer,
+        stemmer,
         access_type: AccessType::Mmap,
     };